==> 4_GenderIndices.R <==
# using the indicator to rank countries and then use our decomposition to illustrate
# the effect

# using data from Gender Equality Index scores, SOURCE: European Institute for Gender Equality (EIGE)
